852631 - system group - update model to raise exception when no groups exist
This commit contains changes so that if the user invokes an action
(e.g. pkg install) on a system group that has no members, they will
receive an error message.

From the UI, this might look like:
  "Install of Packages 'xterm' scheduled for System Group 'group2'
  failed. Reason: System group is empty."

From the CLI, this might look like:
  katello> system_group packages --org myorg --name grp1 --install xterm
  System group is empty.

+++ app/models/system_group.rb
@@ -157,32 +157,38 @@ class SystemGroup < ActiveRecord::Base
   end
 
   def install_packages packages
+    raise Errors::SystemGroupEmptyException if self.systems.empty?
     pulp_job = self.install_package(packages)
     job = save_job(pulp_job, :package_install, :packages, packages)
   end
 
   def uninstall_packages packages
+    raise Errors::SystemGroupEmptyException if self.systems.empty?
     pulp_job = self.uninstall_package(packages)
     job = save_job(pulp_job, :package_remove, :packages, packages)
   end
 
   def update_packages packages=nil
     # if no packages are provided, a full system update will be performed (e.g ''yum update' equivalent)
+    raise Errors::SystemGroupEmptyException if self.systems.empty?
     pulp_job = self.update_package(packages)
     job = save_job(pulp_job, :package_update, :packages, packages)
   end
 
   def install_package_groups groups
+    raise Errors::SystemGroupEmptyException if self.systems.empty?
     pulp_job = self.install_package_group(groups)
     job = save_job(pulp_job, :package_group_install, :groups, groups)
   end
 
   def uninstall_package_groups groups
+    raise Errors::SystemGroupEmptyException if self.systems.empty?
     pulp_job = self.uninstall_package_group(groups)
     job = save_job(pulp_job, :package_group_remove, :groups, groups)
   end
 
   def install_errata errata_ids
+    raise Errors::SystemGroupEmptyException if self.systems.empty?
     pulp_job = self.install_consumer_errata(errata_ids)
     job = save_job(pulp_job, :errata_install, :errata_ids, errata_ids)
   end
